refactor module importing
