+++ src/renderer/ducks/validator/sagas.ts
@@ -19,7 +19,6 @@ import {
 } from "redux-saga/effects";
 import {CGAccount} from "../../models/account";
 import {deleteKeystore, saveValidatorData} from "../../services/utils/account";
-import {ValidatorLogger} from "../../services/eth2/client/logger";
 import database, {cgDbController} from "../../services/db/api/database";
 import {config as mainnetConfig} from "@chainsafe/lodestar-config/lib/presets/mainnet";
 import {IValidator, IValidatorComplete} from "./slice";
@@ -53,8 +52,6 @@ import {Genesis} from "@chainsafe/lodestar-types";
 import {getAuthAccount} from "../auth/selectors";
 import {getValidator, getValidatorsByBeaconNode, BeaconValidators} from "./selectors";
 import {ValidatorBeaconNodes} from "../../models/validatorBeaconNodes";
-import {CgEth2ApiClient} from "../../services/eth2/client/eth2ApiClient";
-import {getBeaconNodeEth2ApiClient, readBeaconChainNetwork} from "../../services/eth2/client";
 import {INetworkConfig} from "../../services/interfaces";
 import {getValidatorBalance} from "../../services/utils/validator";
 import {getValidatorStatus} from "../../services/utils/getValidatorStatus";
@@ -74,6 +71,12 @@ import {Interchange} from "@chainsafe/lodestar-validator/lib/slashingProtection/
 import {createNotification} from "../notification/actions";
 import {Level} from "../../components/Notification/NotificationEnums";
 import {setInitialValidators, setLoadingValidator} from "../settings/actions";
+import {
+    CgEth2ApiClient,
+    getBeaconNodeEth2ApiClient,
+    readBeaconChainNetwork,
+    ValidatorLogger,
+} from "../../services/eth2/client/module";
 
 interface IValidatorServices {
     [validatorAddress: string]: Validator;
@@ -180,7 +183,7 @@ function* startService(
     | Promise<INetworkConfig | null>
     | Promise<Genesis | null>
     | RaceEffect<TakeEffect>
-    | Promise<CgEth2ApiClient>
+    | Promise<typeof CgEth2ApiClient>
     | CancelEffect,
     void,
     IValidatorComplete &
@@ -188,7 +191,7 @@ function* startService(
         [ReturnType<typeof stopActiveValidatorService> | undefined, ReturnType<typeof setValidatorStatus>] &
         (INetworkConfig | null) &
         (Genesis | null) &
-        CgEth2ApiClient &
+        typeof CgEth2ApiClient &
         boolean
 > {
     try {
