+++ src/renderer/ducks/validator/slice.ts
@@ -1,9 +1,9 @@
 import {createSlice, PayloadAction} from "@reduxjs/toolkit";
-import {ValidatorLogger} from "../../services/eth2/client/logger";
 import {ICGKeystore} from "../../services/keystore";
 import {ValidatorStatus} from "../../constants/validatorStatus";
 import {toHexString} from "@chainsafe/ssz";
 import {ValidatorResponse} from "@chainsafe/lodestar-types";
+import {ValidatorLogger} from "../../services/eth2/client/module";
 
 export interface IValidator {
     name: string;
