+++ src/renderer/services/eth2/client/lighthouse/index.ts
@@ -1,3 +1,3 @@
-import {CgEth2ApiClient} from "../eth2ApiClient";
+import {CgEth2ApiClient} from "../module";
 
 export class CgLighthouseEth2Api extends CgEth2ApiClient {}
