+++ src/renderer/services/eth2/client/module.ts
@@ -1,0 +1,16 @@
+// re-export to include modules at specific order
+//
+// eslint-disable-next-line max-len
+// https://medium.com/visual-development/how-to-fix-nasty-circular-dependency-issues-once-and-for-all-in-javascript-typescript-a04c987cf0de
+
+// important order
+export * from "./eth2ApiClient";
+export * from "./lighthouse";
+export * from "./teku";
+
+// as you wish
+export * from "./utils";
+export * from "./logger";
+
+// do not import! idk why but broke everything
+// export * from "./defaults";
