+++ src/renderer/services/eth2/client/utils.ts
@@ -2,9 +2,7 @@ import {INetworkConfig} from "../../interfaces";
 import {getNetworkConfig, getNetworkConfigByGenesisVersion} from "../networks";
 import {ICgEth2ApiClient} from "./interface";
 import {HttpClient} from "../../api";
-import {CgEth2ApiClient} from "./eth2ApiClient";
-import {CgLighthouseEth2Api} from "./lighthouse";
-import {CgTekuEth2Api} from "./teku";
+import {CgLighthouseEth2Api, CgTekuEth2Api, CgEth2ApiClient} from "./module";
 
 export function getEth2ApiClient(url: string, network: string): ICgEth2ApiClient | undefined {
     const networkConfig = getNetworkConfig(network);
@@ -28,7 +26,7 @@ export async function readBeaconChainNetwork(url: string): Promise<INetworkConfi
     }
 }
 
-export async function getBeaconNodeEth2ApiClient(beaconNodeUrl: string): Promise<CgEth2ApiClient> {
+export async function getBeaconNodeEth2ApiClient(beaconNodeUrl: string): Promise<typeof CgEth2ApiClient> {
     const response = await new HttpClient(beaconNodeUrl).get<{data: {version: string}}>(`/eth/v1/node/version`);
     const version = response.data.version.toLowerCase();
